str2int now returns int64

+++ include/carbon/types.h
@@ -6,13 +6,15 @@
 /*   By: ancoulon <ancoulon@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/04 10:48:41 by ancoulon          #+#    #+#             */
-/*   Updated: 2021/03/04 10:50:24 by ancoulon         ###   ########.fr       */
+/*   Updated: 2021/05/28 13:20:40 by ancoulon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef TYPES_H
 # define TYPES_H
 
-int		types_str2int(char *s);
+#include <stdint.h>
+
+int64_t		types_str2int(char *s);
 
 #endif
